requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.7.0
attrs==23.2.0
bcrypt==4.1.3
cffi==1.16.0
cryptography==42.0.8
discord.py==2.3.2
dynaconf==3.2.0
frozenlist==1.4.1
idna==3.7
iso8601==1.1.0
loguru==0.7.2
multidict==6.0.5
paramiko==3.4.0
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
PyNaCl==1.5.0
pypika-tortoise==0.1.6
pytz==2024.1
tortoise-orm==0.21.2
typing_extensions==4.12.0
yarl==1.9.4
pycryptodome==3.20.0
